예상매출 ranking api 개발(#DF4-64) by소영

+++ fp_ml_web/src/main/java/com/springboot/fp_ml_web/data/dto/FilterSelectionDto.java
@@ -30,8 +30,8 @@ public class FilterSelectionDto {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class RentFee {
-        private Integer min;
-        private Integer max;
+        private Double min;
+        private Double max;
     }
 
     @Getter
@@ -39,8 +39,8 @@ public class FilterSelectionDto {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class RentArea {
-        private Float min;
-        private Float max;
+        private Double min;
+        private Double max;
     }
 
 }
